chore: fix types
