+++ src/components/game-screen/InProgressGame.tsx
@@ -12,10 +12,10 @@ import { createClientQuestionStrategy } from "~/lib/games/question-types/client_
 import { type NotWaitingForPlayersGameState } from "~/lib/games/types"
 import { createDefaultLayout, createDefaultMobileLayout } from "~/lib/surfaces/panels/layouts"
 import PanelSkeleton from "~/lib/surfaces/panels/PanelSkeleton"
+import CodeRunning from "./CodeRunning"
 import CodeRunningFooter from "./CodeRunningFooter"
 import { CodeViewImpl } from "./CodeView"
 import MultiSelectPanel from "./MultiSelectPanel"
-import RunCodePanel from "./RunCodePanel"
 
 export const MOBILE_VIEWPORT = "(max-width: 640px)"
 
@@ -37,7 +37,7 @@ function useViews(props: {
     const CodeRunningViewImpl = {
       key: "run-code",
       className: "p-4",
-      component: <RunCodePanel />,
+      component: <CodeRunning questionStrategy={props.questionStrategy} />,
       footer: <CodeRunningFooter />,
       footerClassName: "px-4 pb-4 pt-2",
     }
