--- src/components/game-screen/RunCodePanel.tsx
@@ -1,7 +1,0 @@
-import React from "react"
-
-import CodeRunning from "./CodeRunning"
-
-export default function RunCodePanel() {
-  return <CodeRunning />
-}
