+++ src/lib/games/question-types/picture/client.tsx
@@ -1,10 +1,16 @@
 import { invariant } from "@epic-web/invariant"
 
-import { type FullQuestion } from "../../types"
+import type { FullQuestion } from "../../types"
+import type { ClientQuestionStrategy } from "../base"
+import { PictureQuestionResults } from "~/components/game-screen/PictureQuestionResults"
+import { type QuestionTestState } from "../../types"
 import { BaseQuestionStrategy } from "../base"
 import { PictureQuestionConfig } from "./config"
 
-export class PictureQuestionStrategy extends BaseQuestionStrategy {
+export class PictureQuestionStrategy
+  extends BaseQuestionStrategy
+  implements ClientQuestionStrategy
+{
   private readonly question: FullQuestion & {
     pictureQuestion: NonNullable<FullQuestion["pictureQuestion"]>
   }
@@ -25,5 +31,9 @@ export class PictureQuestionStrategy extends BaseQuestionStrategy {
 
   get preview() {
     return <div>Picture question</div>
+  }
+
+  results(testState: QuestionTestState | null) {
+    return <PictureQuestionResults question={this.question} testState={testState} />
   }
 }
